feat(tabular): enhance balanced widths functionality and add examples
- Implemented additional tests for balanced widths, ensuring proportional distribution across columns.
- Updated README.md with detailed examples demonstrating the new balanced widths feature.
- Introduced a new example file showcasing various scenarios for balanced widths in table layouts.
- Refactored existing tests to improve coverage and validate the interaction between balanced widths and other table properties.
