+++ packages/tabular/LICENSE.md
@@ -23,14 +23,11 @@ SOFTWARE.
 <!-- DEPENDENCIES -->
 
 # Licenses of bundled dependencies
-
 The published @visulima/tabular artifact additionally contains code with the following licenses:
 MIT
 
 # Bundled dependencies:
-
 ## @visulima/string
-
 License: MIT
 By: Daniel Bannert
 Repository: git+https://github.com/visulima/visulima.git
@@ -57,15 +54,14 @@ Repository: git+https://github.com/visulima/visulima.git
 > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > SOFTWARE.
 >
-> # Licenses of bundled dependencies
 >
+>
+> # Licenses of bundled dependencies
 > The published @visulima/string artifact additionally contains code with the following licenses:
 > MIT
 >
 > # Bundled dependencies:
->
 > ## emoji-regex-xs
->
 > License: MIT
 > By: Steven Levithan
 > Repository: git+https://github.com/slevithan/emoji-regex-xs.git
@@ -92,10 +88,9 @@ Repository: git+https://github.com/visulima/visulima.git
 > > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > > SOFTWARE.
 >
-> ---
+> ---------------------------------------
 >
 > ## fastest-levenshtein
->
 > License: MIT
 > By: Kasper U. Weihe
 > Repository: git+https://github.com/ka-weihe/fastest-levenshtein.git
@@ -122,10 +117,9 @@ Repository: git+https://github.com/visulima/visulima.git
 > > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > > SOFTWARE.
 >
-> ---
+> ---------------------------------------
 >
 > ## get-east-asian-width
->
 > License: MIT
 > By: Sindre Sorhus
 > Repository: sindresorhus/get-east-asian-width
@@ -140,15 +134,16 @@ Repository: git+https://github.com/visulima/visulima.git
 > >
 > > THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 >
-> # Licenses of bundled types
 >
+>
+>
+>
+> # Licenses of bundled types
 > The published @visulima/string artifact additionally contains code with the following licenses:
 > MIT
 >
 > # Bundled types:
->
 > ## fastest-levenshtein
->
 > License: MIT
 > By: Kasper U. Weihe
 > Repository: git+https://github.com/ka-weihe/fastest-levenshtein.git
@@ -175,10 +170,9 @@ Repository: git+https://github.com/visulima/visulima.git
 > > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > > SOFTWARE.
 
----
+---------------------------------------
 
 ## terminal-size
-
 License: MIT
 By: Sindre Sorhus
 Repository: sindresorhus/terminal-size
@@ -198,14 +192,11 @@ Repository: sindresorhus/terminal-size
 <!-- TYPE_DEPENDENCIES -->
 
 # Licenses of bundled types
-
 The published @visulima/tabular artifact additionally contains code with the following licenses:
 MIT
 
 # Bundled types:
-
 ## @visulima/string
-
 License: MIT
 By: Daniel Bannert
 Repository: git+https://github.com/visulima/visulima.git
@@ -235,14 +226,11 @@ Repository: git+https://github.com/visulima/visulima.git
 > <!-- DEPENDENCIES -->
 >
 > # Licenses of bundled dependencies
->
 > The published @visulima/string artifact additionally contains code with the following licenses:
 > MIT
 >
 > # Bundled dependencies:
->
 > ## emoji-regex-xs
->
 > License: MIT
 > By: Steven Levithan
 > Repository: git+https://github.com/slevithan/emoji-regex-xs.git
@@ -269,10 +257,9 @@ Repository: git+https://github.com/visulima/visulima.git
 > > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > > SOFTWARE.
 >
-> ---
+> ---------------------------------------
 >
 > ## fastest-levenshtein
->
 > License: MIT
 > By: Kasper U. Weihe
 > Repository: git+https://github.com/ka-weihe/fastest-levenshtein.git
@@ -299,10 +286,9 @@ Repository: git+https://github.com/visulima/visulima.git
 > > OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 > > SOFTWARE.
 >
-> ---
+> ---------------------------------------
 >
 > ## get-east-asian-width
->
 > License: MIT
 > By: Sindre Sorhus
 > Repository: sindresorhus/get-east-asian-width
@@ -319,15 +305,14 @@ Repository: git+https://github.com/visulima/visulima.git
 >
 > <!-- /DEPENDENCIES -->
 >
-> # Licenses of bundled types
 >
+>
+> # Licenses of bundled types
 > The published @visulima/string artifact additionally contains code with the following licenses:
 > MIT
 >
 > # Bundled types:
->
 > ## fastest-levenshtein
->
 > License: MIT
 > By: Kasper U. Weihe
 > Repository: git+https://github.com/ka-weihe/fastest-levenshtein.git
