+++ packages/tabular/__tests__/table-core.test.ts
@@ -484,202 +484,6 @@ describe("table core functionality", () => {
     });
 });
 
-describe("balancedWidths", () => {
-    it("should distribute available width proportionally across columns when balancedWidths is enabled", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 30,
-        });
-
-        table.addRow(["Short", "Medium length text", "Very long column content here"]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌─────┬──────────┬───────────┐
-            │Short│Medium le…│Very long …│
-            └─────┴──────────┴───────────┘"
-        `);
-    });
-
-    it("should respect minimum column widths while balancing", () => {
-        expect.assertions(2);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 25,
-        });
-
-        table.addRow(["A", "B", "C"]);
-
-        // Should distribute the 25 character width more evenly across 3 columns
-        const output = table.toString();
-        const lines = output.split("\n");
-        const topBorder = lines[0] ?? "";
-
-        // With balanced widths, columns should be more evenly distributed
-        // The total width should be close to 25 (accounting for borders)
-        expect(topBorder.length).toBeGreaterThan(20);
-        expect(topBorder.length).toBeLessThanOrEqual(25 + 2); // Allow small buffer
-    });
-
-    it("should balance widths with word wrapping enabled", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 40,
-            wordWrap: true,
-        });
-
-        table.addRow(["Short text", "This is a longer text that might wrap to multiple lines", "Medium length content here"]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌────────────┬────────────┬────────────┐
-            │Short text  │This is a   │Medium      │
-            │            │longer text │length      │
-            │            │that might  │content here│
-            │            │wrap to     │            │
-            │            │multiple    │            │
-            │            │lines       │            │
-            └────────────┴────────────┴────────────┘"
-        `);
-    });
-
-    it("should work with different padding values", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 2, paddingRight: 1 },
-            terminalWidth: 35,
-        });
-
-        table.addRow(["Col1", "Column 2", "Third Column"]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌──────────┬─────────┬────────────┐
-            │  Col1    │  Colum… │  Third Co… │
-            └──────────┴─────────┴────────────┘"
-        `);
-    });
-
-    it("should handle empty content columns", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 30,
-        });
-
-        table.addRow(["", "Some content", ""]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌─────────┬──────────┬───────┐
-            │         │Some cont…│       │
-            └─────────┴──────────┴───────┘"
-        `);
-    });
-
-    it("should balance widths with fixed column widths taking precedence", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            columnWidths: [5, undefined, 8],
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 30,
-        });
-
-        table.addRow(["A", "Very long content here", "B"]);
-
-        // Fixed widths should be respected, only undefined widths should be balanced
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌─────┬─────────────┬────────┐
-            │A    │Very long co…│B       │
-            └─────┴─────────────┴────────┘"
-        `);
-    });
-
-    it("should consider word wrapping when calculating balanced widths", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 50,
-            wordWrap: true, // Enable global word wrapping
-        });
-
-        table.addRow([
-            "Short", // Short content gets full width (no need to wrap)
-            "This is a very long sentence that can wrap to multiple lines", // Long content can be narrower since it can wrap
-            { content: "Non-wrappable very long content here", wordWrap: false }, // Cannot wrap, so gets reasonable width based on longest word + buffer
-        ]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌──────────────┬────────────┬────────────────────┐
-            │Short         │This is a   │Non-wrappable very …│
-            │              │very long   │                    │
-            │              │sentence    │                    │
-            │              │that can    │                    │
-            │              │wrap to     │                    │
-            │              │multiple    │                    │
-            │              │lines       │                    │
-            └──────────────┴────────────┴────────────────────┘"
-        `);
-    });
-
-    it("should allow cells to override word wrap behavior for balanced widths", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 45,
-            wordWrap: true, // Global word wrap enabled
-        });
-
-        table.addRow([
-            "Short text",
-            { content: "Long content that can wrap", wordWrap: false }, // Explicitly disable wrapping
-            "Another wrappable text here",
-        ]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌─────────────┬─────────────────┬───────────┐
-            │Short text   │Long content tha…│Another    │
-            │             │                 │wrappable  │
-            │             │                 │text here  │
-            └─────────────┴─────────────────┴───────────┘"
-        `);
-    });
-
-    it("should work with balanced widths and default terminal width", () => {
-        expect.assertions(1);
-
-        const table = new Table({
-            balancedWidths: true,
-            style: { paddingLeft: 0, paddingRight: 0 },
-            terminalWidth: 80,
-            wordWrap: true,
-            // No terminalWidth specified - uses default
-        });
-
-        table.addRow(["Column 1 content", "Column 2 with longer text"]);
-
-        expect(table.toString()).toMatchInlineSnapshot(`
-            "┌───────────────────────────────────────┬──────────────────────────────────────┐
-            │Column 1 content                       │Column 2 with longer text             │
-            └───────────────────────────────────────┴──────────────────────────────────────┘"
-        `);
-    });
-});
-
 describe("error handling and edge cases", () => {
     it("should throw if addRow receives non-array input", () => {
         expect.assertions(1);
